requirements.txt
streamlit
requests
python-dotenv
pytest
python-dateutil
sqlalchemy
langchain-openai
faiss-cpu
langchain
langchain-community
langchain-core